pyproject.toml
[tool.poetry]
name = "funcnodes-sec"
version = "0.1.0"
description = ""
authors = ["Kourosh Rezaei <kourosh.rezaei@linkdlab.de>"]
readme = "README.md"
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]


[tool.poetry.urls]
homepage = "https://github.com/Linkdlab/funcnodes_sec"
source = "https://github.com/Linkdlab/funcnodes_sec"
tracker = "https://github.com/Linkdlab/funcnodes_sec/issues"
download = "https://pypi.org/project/funcnodes-sec/#files"

[tool.poetry.dependencies]
python = ">=3.11"
funcnodes = "*"
funcnodes-span = "*"
funcnodes-pandas = "*"


[tool.poetry.group.dev.dependencies]
pytest = "*"
pre-commit = "*"
funcnodes-module = "^0.1.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.plugins."funcnodes.module"]
module = "funcnodes_sec"
shelf = "funcnodes_sec:NODE_SHELF"
